Remove view model
